+++ Kernel/cpan-lib/Linux/Distribution.pm
@@ -10,7 +10,7 @@ our @ISA = qw(Exporter);
 
 our @EXPORT_OK = qw( distribution_name distribution_version );
 
-our $VERSION = '0.21';
+our $VERSION = '0.23';
 
 our $release_files_directory='/etc';
 our $standard_release_file = 'lsb-release';
@@ -43,6 +43,8 @@ our %release_files = (
     'libranet_version'      => 'libranet',
     'va-release'            => 'va-linux',
     'pardus-release'        => 'pardus',
+    'system-release'        => 'amazon',
+    'CloudLinux-release'    => 'CloudLinux',
 );
 
 our %version_match = (
@@ -55,8 +57,10 @@ our %version_match = (
     'oracle enterprise linux' => 'Enterprise Linux Server release (.+) \(',
     'slackware'             => '^Slackware (.+)$',
     'pardus'                => '^Pardus (.+)$',
-    'centos'                => '^CentOS(?: Linux)? release (.+)(?:\s\(Final\))',
+    'centos'                => '^CentOS(?: Linux)? release (.+) \(',
     'scientific'            => '^Scientific Linux release (.+) \(',
+    'amazon'                => 'Amazon Linux AMI release (.+)$',
+    'CloudLinux'            => 'CloudLinux Server release (\S+)'
 );
 
 
@@ -85,7 +89,7 @@ sub distribution_name {
         return $distro if ($distro);
     }
 
-    foreach (qw(enterprise-release fedora-release)) {
+    foreach (qw(enterprise-release fedora-release CloudLinux-release)) {
         if (-f "$release_files_directory/$_" && !-l "$release_files_directory/$_"){
             if (-f "$release_files_directory/$_" && !-l "$release_files_directory/$_"){
                 $self->{'DISTRIB_ID'} = $release_files{$_};
@@ -138,7 +142,7 @@ sub _get_lsb_info {
     my $tmp = $self->{'release_file'};
     if ( -r "$release_files_directory/" . $standard_release_file ) {
         $self->{'release_file'} = $standard_release_file;
-        $self->{'pattern'} = $field . '=(.+)';
+        $self->{'pattern'} = $field . '=["]?([^"]+)["]?';
         my $info = $self->_get_file_info();
         if ($info){
             $self->{$field} = $info;
@@ -198,9 +202,9 @@ Linux::Distribution - Perl extension to detect on which Linux distribution we ar
 
 This is a simple module that tries to guess on what linux distribution we are running by looking for release's files in /etc.  It now looks for 'lsb-release' first as that should be the most correct and adds ubuntu support.  Secondly, it will look for the distro specific files.
 
-It currently recognizes slackware, debian, suse, fedora, redhat, turbolinux, yellowdog, knoppix, mandrake, conectiva, immunix, tinysofa, va-linux, trustix, adamantix, yoper, arch-linux, libranet, gentoo, ubuntu, scientific, oracle enterprise linux and redflag.
+It currently recognizes slackware, debian, suse, fedora, redhat, turbolinux, yellowdog, knoppix, mandrake, conectiva, immunix, tinysofa, va-linux, trustix, adamantix, yoper, arch-linux, libranet, gentoo, ubuntu, scientific, oracle enterprise linux, amazon linux and redflag.
 
-It has function to get the version for debian, suse, fedora, redhat, gentoo, slackware, scientific, oracle enterprise linux, redflag and ubuntu(lsb). People running unsupported distro's are greatly encouraged to submit patches :-)
+It has function to get the version for debian, suse, fedora, redhat, gentoo, slackware, scientific, oracle enterprise linux, amazon linux, redflag and ubuntu(lsb). People running unsupported distro's are greatly encouraged to submit patches :-)
 
 =head2 EXPORT